requirements.txt
pandas==1.3.4
Pillow==9.1.1
scikit_learn==1.1.1
streamlit==1.10.0
